ubuntu 64bit running with uintptr_t

+++ src/bitmap.c
@@ -27,6 +27,7 @@
 #include <string.h>
 #include <noftypes.h>
 #include <bitmap.h>
+#include <stdint.h>
 
 void bmp_clear(const bitmap_t *bitmap, uint8 color)
 {
@@ -61,7 +62,7 @@ static bitmap_t *_make_bitmap(uint8 *data_addr, bool hw, int width,
    if (false == bitmap->hardware)
    {
       bitmap->pitch = (bitmap->pitch + 3) & ~3;
-      bitmap->line[0] = (uint8 *) (((uint32) bitmap->data + overdraw + 3) & ~3);
+      bitmap->line[0] = (uint8 *) (((uintptr_t) bitmap->data + overdraw + 3) & ~3);
    }
    else
    {
